address feedback to move staking status icon to a component
