+++ src/pages/DelegatoryValidator/Components/MyDepositsStatusTooltip.tsx
@@ -10,10 +10,10 @@ import {
 import React from "react";
 import TableTooltip from "../../../components/Table/TableTooltip";
 import TooltipTypography from "../../../components/TooltipTypography";
-import {StakingStatusStepInterface} from "../MyDepositsSection";
+import {StakingStatusStepInterface} from "./StakingStatusIcon";
 
 type MyDepositsSectionProps = {
-  steps: [StakingStatusStepInterface];
+  steps: StakingStatusStepInterface[];
 };
 
 // TODO(jill): refactor step icon color scheme to override the default
